+++ src/ClientConfig.php
@@ -51,9 +51,6 @@ class ClientConfig
             // wait this long (seconds) before trying the next server in the list
             'server-poll-timeout 10',
 
-            // CRYPTO (DATA CHANNEL)
-            'auth SHA256',
-
             // CRYPTO (CONTROL CHANNEL)
             // @see RFC 7525
             // @see https://bettercrypto.org
@@ -84,6 +81,7 @@ class ClientConfig
         }
 
         if ($profileConfig['tlsCrypt']) {
+            // >= 2.4
             $clientConfig = array_merge(
                 $clientConfig,
                 [
@@ -92,10 +90,10 @@ class ClientConfig
                     '</tls-crypt>',
                 ]
             );
-            // if tls-crypt is enabled, we need OpenVPN >= 2.4 anyway, so also
-            // bump cipher
-            $clientConfig[] = 'cipher AES-256-GCM';
+        // no need to specify --auth or --cipher for 2.4 clients with
+            // tls-crypt
         } else {
+            // < 2.4
             $clientConfig = array_merge(
                 $clientConfig,
                 [
@@ -105,10 +103,8 @@ class ClientConfig
                     '</tls-auth>',
                 ]
             );
-            // if tls-crypt is not enabled, we still want to maintain OpenVPN
-            // 2.3 compat, so allow CBC as well, NCP will override this for
-            // 2.4 clients anyway
             $clientConfig[] = 'cipher AES-256-CBC';
+            $clientConfig[] = 'auth SHA256';
         }
 
         // --comp-lzo