+++ src/OAuthClientInfo.php
@@ -62,13 +62,11 @@ class OAuthClientInfo
             // iOS
             'org.eduvpn.app.ios' => [
                 'redirect_uri_list' => [
-                    'org.eduvpn.app:/api/callback',
-                    // https://github.com/eduvpn/ios/issues/23
                     'https://ios.app.eduvpn.org/auth/app/redirect/',
                     'https://ios.app.eduvpn.org/auth/app/redirect/development/',
                 ],
                 'display_name' => 'eduVPN for iOS',
-                'require_approval' => true,
+                'require_approval' => false,
             ],
             // macOS
             'org.eduvpn.app.macos' => [