+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.5.3 (...)
 - switch to `UserInfo` class
+- no longer need --auth and --cipher for OpenVPN >= 2.4 clients
 
 ## 1.5.2 (2018-02-28)
 - make sure chosen userId does not exist yet when registering a new account